uint to uint64 because Contains is not smart enough unfortunately

+++ tracker_windows_test.go
@@ -47,8 +47,8 @@ func TestWindowsProcessTracker(t *testing.T) {
 			info, err := QueryInformationJobObjectProcessIdList(tracker.job.handle)
 			assert.NoError(t, err)
 			assert.Equal(t, 2, int(info.NumberOfAssignedProcesses))
-			assert.Contains(t, info.ProcessIdList, uint(proc1.Info(ctx).PID))
-			assert.Contains(t, info.ProcessIdList, uint(proc2.Info(ctx).PID))
+			assert.Contains(t, info.ProcessIdList, uint64(proc1.Info(ctx).PID))
+			assert.Contains(t, info.ProcessIdList, uint64(proc2.Info(ctx).PID))
 		},
 		"AddedProcessIsTerminatedOnCleanup": func(ctx context.Context, t *testing.T, tracker *windowsProcessTracker, opts *CreateOptions) {
 			proc, err := newBasicProcess(ctx, opts)
@@ -59,7 +59,7 @@ func TestWindowsProcessTracker(t *testing.T) {
 			info, err := QueryInformationJobObjectProcessIdList(tracker.job.handle)
 			assert.NoError(t, err)
 			assert.Equal(t, 1, int(info.NumberOfAssignedProcesses))
-			assert.Contains(t, info.ProcessIdList, uint(proc.Info(ctx).PID))
+			assert.Contains(t, info.ProcessIdList, uint64(proc.Info(ctx).PID))
 
 			assert.NoError(t, tracker.Cleanup())
 
@@ -81,7 +81,7 @@ func TestWindowsProcessTracker(t *testing.T) {
 			info, err := QueryInformationJobObjectProcessIdList(tracker.job.handle)
 			assert.NoError(t, err)
 			assert.Equal(t, 1, int(info.NumberOfAssignedProcesses))
-			assert.Contains(t, info.ProcessIdList, uint(proc.Info(ctx).PID))
+			assert.Contains(t, info.ProcessIdList, uint64(proc.Info(ctx).PID))
 
 			assert.NoError(t, tracker.Cleanup())
 			assert.NoError(t, tracker.Cleanup())
